Move common S3 logic into a shared module
