+++ lib/manageiq/rpm_build.rb
@@ -7,11 +7,14 @@ class String
   include Term::ANSIColor
 end
 
+require 'manageiq/rpm_build/s3_common'
+
 require 'manageiq/rpm_build/helper'
 require 'manageiq/rpm_build/build_copr'
 require 'manageiq/rpm_build/build_uploader'
 require 'manageiq/rpm_build/generate_gemset'
 require 'manageiq/rpm_build/generate_tar_files'
+require 'manageiq/rpm_build/rpm_repo'
 require 'manageiq/rpm_build/setup_source_repos'
 
 module ManageIQ
