+++ lib/manageiq/rpm_build/s3_common.rb
@@ -1,0 +1,46 @@
+module ManageIQ
+  module RPMBuild
+    module S3Common
+      private
+
+      def client
+        @client ||= begin
+          require 'aws-sdk-s3'
+          Aws::S3::Client.new(
+            :access_key_id     => OPTIONS.rpm_repository.s3_api.access_key,
+            :secret_access_key => OPTIONS.rpm_repository.s3_api.secret_key,
+            :region            => 'us-east-1',
+            :endpoint          => "https://#{OPTIONS.rpm_repository.s3_api.endpoint}"
+          )
+        end
+      end
+
+      def remote_etag(key)
+        client.head_object(
+          :bucket => OPTIONS.rpm_repository.s3_api.bucket,
+          :key    => key
+        )[:etag].tr("\\\"", "")
+      rescue Aws::S3::Errors::NotFound
+      end
+
+      def upload_file(source, destination)
+        require 'digest'
+        md5sum = Digest::MD5.file(source).hexdigest
+
+        if md5sum == remote_etag(destination)
+          puts "  Skipping existing file: #{destination}"
+        else
+          puts "  Uploading: #{destination}"
+          File.open(source, 'rb') do |content|
+            client.put_object(
+              :bucket => OPTIONS.rpm_repository.s3_api.bucket,
+              :key    => destination,
+              :body   => content,
+              :acl    => 'public-read'
+            )
+          end
+        end
+      end
+    end
+  end
+end
